Don't ignore unknown exceptions raised by get_router.
If an unknown exception is raised on one NS document during
read_routers, it's very likely it's going to be raised again on the
next, and the next... (e.g., TorCtlClosed).

Instead of printing out the traceback and continuing, allow it to
unwind the stack and let the caller handle it.